requirements.txt
Flask==0.10
Flask-Bootstrap==3.2.0.2
simplejson==3.6.0
Pillow==2.5.1
flask-cors==3.0.3
